compute correlation coefficient

+++ cla_parse.cpp
@@ -9,6 +9,7 @@
 
 #include "./include/cla_parse.hpp"
 
+
 // parse command line arguments
 int
 parse_arguments(
@@ -17,9 +18,9 @@ parse_arguments(
     std::string* image_filename,
     std::string* template_filename,
     std::string* warp_filename,
-    bool* manual,
-    float* epsilon,
-    int* motion_type,
+    bool*        manual,
+    double*      epsilon,
+    int*         motion_type,
     std::string* output_warp,
     std::string* warp_img_filename
 ) {
